Deprecate seqSpine
The `seqSpine` function was useful back in the day when all sorts of
operations were written in a way that allowed thunks to accumulate along
the spine. Now, we carefully force as much of the spine as the amortized
bounds allow, which eliminates this problem in general. The only
operations that still accumulate thunks along the spine are monotonic
maps, and forcing the spines of those just leads to the creation of more
thunks at the leaves—it's not really useful.