Make Prio queues more compact
Store the value associated with each key as its rightmost child,
which saves one word per element.

As a result, the binomial trees must become lazy, which should be
good for maps and lazy traversals. The down side is that we will
need tag checks to know that we have realized `Succ` constructors.

Benchmarking suggests this implementation is substantially faster
than the previous one.